+++ apps/worker/package.json
@@ -21,6 +21,7 @@
     "drizzle-orm": "1.0.0-beta.22",
     "ts-pattern": "^5.9.0",
     "tsx": "^4.22.4",
+    "uuid": "^14.0.1",
     "zod": "^4.4.3"
   },
   "devDependencies": {
